chore: Update new-hire component to calculate and display days to go
The new-hire component has been updated to calculate and display the number of days remaining until the employee's first day. This is achieved by retrieving the employee's date of joining from the API and calculating the time difference between the current date and the first day. The days to go value is then displayed in the component template.